refactor(ui): split configuration fixtures and app-tab tests
