+++ crates/mesh-llm-ui/src/features/app-tabs/configuration-fixtures.ts
@@ -1,0 +1,213 @@
+import type { ConfigAssign, ConfigModel, ConfigNode } from '@/features/app-tabs/types'
+
+export const CFG_NODES: ConfigNode[] = [
+  {
+    id: 'node-a',
+    hostname: 'carrack',
+    region: 'tor-1',
+    status: 'online',
+    cpu: 'Ryzen 7950X',
+    ramGB: 128,
+    placement: 'separate',
+    gpus: [
+      { idx: 0, name: 'RTX 5090', totalGB: 34.2, reservedGB: 0.9 },
+      { idx: 1, name: 'RTX 6000 Pro', totalGB: 48.0, reservedGB: 1.1 },
+      { idx: 2, name: 'RTX 6000 Pro', totalGB: 48.0, reservedGB: 1.1 },
+      { idx: 3, name: 'RTX 6000 Pro', totalGB: 48.0, reservedGB: 1.1 },
+      { idx: 4, name: 'RTX 6000 Pro', totalGB: 48.0, reservedGB: 1.1 },
+      { idx: 5, name: 'RTX 6000 Pro', totalGB: 48.0, reservedGB: 1.1 },
+      { idx: 6, name: 'RTX 6000 Pro', totalGB: 48.0, reservedGB: 1.1 },
+      { idx: 7, name: 'RTX 3080', totalGB: 10.7, reservedGB: 0.6 }
+    ]
+  },
+  {
+    id: 'node-b',
+    hostname: 'perseus.local',
+    region: 'unified',
+    status: 'online',
+    cpu: 'Apple M4 Pro',
+    ramGB: 64,
+    placement: 'pooled',
+    memoryTopology: 'unified',
+    gpus: [{ idx: 0, name: 'unified memory', totalGB: 51.5, reservedGB: 1.5 }]
+  },
+  {
+    id: 'node-c',
+    hostname: 'triton.lab',
+    region: 'sfo-1',
+    status: 'offline',
+    cpu: 'Xeon W',
+    ramGB: 96,
+    placement: 'separate',
+    gpus: [{ idx: 0, name: 'RTX 4090', totalGB: 24.0, reservedGB: 0.8 }]
+  }
+]
+export const CFG_CATALOG: ConfigModel[] = [
+  {
+    id: 'glm47',
+    name: 'GLM-4.7-Flash-Q4_K_M',
+    family: 'GLM',
+    familyColor: 'family-0',
+    paramsB: 4.7,
+    paramsLabel: '~70B',
+    quant: 'Q4_K_M',
+    sizeGB: 18.5,
+    diskGB: 19,
+    ctxMaxK: 128,
+    ctxPerGB: 0.017,
+    layers: 80,
+    heads: 40,
+    embed: 8192,
+    tokenizer: 'glm',
+    moe: false,
+    vision: false,
+    tags: ['chat']
+  },
+  {
+    id: 'llama70',
+    name: 'Llama-3.3-70B-Q4_K_M',
+    family: 'Llama',
+    familyColor: 'family-1',
+    paramsB: 70,
+    paramsLabel: '70B',
+    quant: 'Q4_K_M',
+    sizeGB: 40.3,
+    diskGB: 46,
+    ctxMaxK: 256,
+    ctxPerGB: 0.019,
+    layers: 80,
+    heads: 64,
+    embed: 8192,
+    tokenizer: 'llama',
+    moe: false,
+    vision: false,
+    tags: ['chat', 'tools']
+  },
+  {
+    id: 'qwen27',
+    name: 'Qwen3.5-27B-Q4_K_M',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 27,
+    paramsLabel: '27B',
+    quant: 'Q4_K_M',
+    sizeGB: 17.4,
+    diskGB: 19,
+    ctxMaxK: 64,
+    ctxPerGB: 0.022,
+    layers: 64,
+    heads: 40,
+    embed: 5120,
+    tokenizer: 'qwen',
+    moe: false,
+    vision: false,
+    tags: ['chat']
+  },
+  {
+    id: 'qwen4',
+    name: 'Qwen3-4B-Q4_K_M',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 4,
+    paramsLabel: '4B',
+    quant: 'Q4_K_M',
+    sizeGB: 2.6,
+    diskGB: 3,
+    ctxMaxK: 32,
+    ctxPerGB: 0.018,
+    layers: 36,
+    heads: 28,
+    embed: 3584,
+    tokenizer: 'qwen',
+    moe: false,
+    vision: false,
+    tags: ['chat']
+  },
+  {
+    id: 'qwenud',
+    name: 'Qwen3.5-27B-UD-Q4_K_XL',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 27,
+    paramsLabel: '27B UD',
+    quant: 'Q4_K_XL',
+    sizeGB: 17.8,
+    diskGB: 19,
+    ctxMaxK: 256,
+    ctxPerGB: 0.025,
+    layers: 64,
+    heads: 40,
+    embed: 5120,
+    tokenizer: 'qwen',
+    moe: false,
+    vision: false,
+    tags: ['chat']
+  },
+  {
+    id: 'mixtral',
+    name: 'mixtral-8x22b',
+    family: 'Mixtral',
+    familyColor: 'family-3',
+    paramsB: 176,
+    paramsLabel: '8x22B MoE',
+    quant: 'Q4_K_M',
+    sizeGB: 86,
+    diskGB: 91,
+    ctxMaxK: 64,
+    ctxPerGB: 0.028,
+    layers: 56,
+    heads: 48,
+    embed: 6144,
+    tokenizer: 'mixtral',
+    moe: true,
+    vision: false,
+    tags: ['moe']
+  },
+  {
+    id: 'llava',
+    name: 'llava-next-34b',
+    family: 'LLaVA',
+    familyColor: 'family-4',
+    paramsB: 34,
+    paramsLabel: '34B',
+    quant: 'Q4_K_M',
+    sizeGB: 22,
+    diskGB: 26,
+    ctxMaxK: 32,
+    ctxPerGB: 0.024,
+    layers: 48,
+    heads: 52,
+    embed: 7168,
+    tokenizer: 'llava',
+    moe: false,
+    vision: true,
+    tags: ['vision']
+  },
+  {
+    id: 'phi4',
+    name: 'phi-4-mini',
+    family: 'Phi',
+    familyColor: 'family-6',
+    paramsB: 3.8,
+    paramsLabel: '3.8B',
+    quant: 'Q8_0',
+    sizeGB: 5.2,
+    diskGB: 6.1,
+    ctxMaxK: 32,
+    ctxPerGB: 0.014,
+    layers: 32,
+    heads: 32,
+    embed: 3072,
+    tokenizer: 'phi',
+    moe: false,
+    vision: false,
+    tags: ['small']
+  }
+]
+export const INITIAL_ASSIGNS: ConfigAssign[] = [
+  { id: 'a1', modelId: 'glm47', nodeId: 'node-a', containerIdx: 0, ctx: 16384 },
+  { id: 'a2', modelId: 'llama70', nodeId: 'node-a', containerIdx: 1, ctx: 16384 },
+  { id: 'a3', modelId: 'qwen27', nodeId: 'node-a', containerIdx: 2, ctx: 16384 },
+  { id: 'a4', modelId: 'qwen4', nodeId: 'node-a', containerIdx: 7, ctx: 4096 },
+  { id: 'a5', modelId: 'qwenud', nodeId: 'node-b', containerIdx: 0, ctx: 262144 }
+]
