+++ crates/mesh-llm-ui/src/features/app-tabs/dashboard-fixtures.ts
@@ -1,0 +1,244 @@
+import { createElement } from 'react'
+import { Activity, Cpu, HardDrive, Hash, Network, UserRound, type LucideIcon } from 'lucide-react'
+import type { MeshNode, ModelSummary, Peer, PeerSummary, StatusMetric } from '@/features/app-tabs/types'
+
+const metricIcon = (Icon: LucideIcon) => createElement(Icon, { className: 'size-[11px] shrink-0', 'aria-hidden': true })
+
+export const STATUS_METRICS: StatusMetric[] = [
+  {
+    id: 'node-id',
+    icon: metricIcon(Hash),
+    label: 'Node ID',
+    value: '990232e1c1',
+    variant: 'identity',
+    mono: true,
+    badge: { label: 'Serving', tone: 'good' }
+  },
+  {
+    id: 'owner',
+    icon: metricIcon(UserRound),
+    label: 'Owner',
+    value: 'Unsigned',
+    variant: 'identity',
+    mono: true,
+    badge: { label: 'not cryptographically bound', tone: 'muted' }
+  },
+  { id: 'nodes', icon: metricIcon(Network), label: 'Nodes', value: 3, meta: '1 you · 2 peers' },
+  { id: 'active-models', icon: metricIcon(Cpu), label: 'Active models', value: 6, meta: '2 loaded locally · 4 remote' },
+  {
+    id: 'mesh-vram',
+    icon: metricIcon(HardDrive),
+    label: 'Mesh VRAM',
+    value: '160.5',
+    unit: 'GB',
+    meta: '57% free',
+    sparkline: [12, 14, 11, 16, 13, 15, 17, 14, 18]
+  },
+  {
+    id: 'inflight',
+    icon: metricIcon(Activity),
+    label: 'Inflight',
+    value: 0,
+    meta: '0 requests in flight',
+    sparkline: [4, 8, 5, 12, 6, 14, 7, 9, 5]
+  }
+]
+export const PEERS: Peer[] = [
+  {
+    id: 'p1',
+    shortId: '990232e1c1',
+    hostname: 'carrack',
+    region: 'tor-1',
+    status: 'online',
+    role: 'you',
+    version: '0.64.0',
+    hostedModels: ['Qwen3.6-27B-UD', 'Qwen3.5-4B-UD'],
+    sharePct: 38,
+    latencyMs: 1,
+    vramGB: 61.7,
+    toksPerSec: 36.8,
+    loadPct: 38,
+    hardwareLabel: 'Jetson AGX Orin · 61 GB',
+    ownership: 'Unsigned',
+    owner: 'Unsigned'
+  },
+  {
+    id: 'p2',
+    shortId: 'e5c42cc0ad',
+    hostname: 'lemony-28',
+    region: 'nyc-2',
+    status: 'online',
+    role: 'host',
+    version: '0.64.0',
+    hostedModels: ['Qwen3.6-35B-A3B-UD', 'gemma-4-26B-A4B-it-UD'],
+    sharePct: 31,
+    latencyMs: 1,
+    vramGB: 49.4,
+    toksPerSec: 22.4,
+    loadPct: 31,
+    hardwareLabel: 'Jetson AGX Orin · 61 GB',
+    ownership: 'Unsigned',
+    owner: 'Unsigned'
+  },
+  {
+    id: 'p3',
+    shortId: '7d13fd27b8',
+    hostname: 'lemony-29',
+    region: 'sfo-1',
+    status: 'online',
+    role: 'host',
+    version: '0.64.0',
+    hostedModels: ['Qwen3.5-0.8B-UD', 'Qwen3.5-2B'],
+    sharePct: 31,
+    latencyMs: 1.1,
+    vramGB: 49.4,
+    toksPerSec: 14.2,
+    loadPct: 31,
+    hardwareLabel: 'Mac Studio M2 Ultra · 64 GB',
+    ownership: 'Unsigned',
+    owner: 'Unsigned'
+  }
+]
+export const PEER_SUMMARY: PeerSummary = { total: 3, online: 3, capacity: 'all serving' }
+export const MODELS: ModelSummary[] = [
+  {
+    name: 'gemma-4-26B-A4B-it-UD',
+    fullId: 'gemma-4-26B-A4B-it-UD-Q4_K_XL',
+    family: 'Gemma',
+    familyColor: 'family-5',
+    paramsB: 26,
+    paramsLabel: '26B A4B',
+    quant: 'Q4_K_XL',
+    size: '14.8 GB',
+    sizeGB: 14.8,
+    diskGB: 16,
+    context: '64k',
+    ctxMaxK: 64,
+    ctxPerGB: 0.021,
+    moe: false,
+    vision: false,
+    status: 'warm',
+    tags: ['Text'],
+    nodeCount: 1,
+    activitySummary: '0 requests seen · active 4 min ago'
+  },
+  {
+    name: 'Qwen3.5-0.8B-UD',
+    fullId: 'Qwen3.5-0.8B-UD-Q4_K_XL',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 0.8,
+    paramsLabel: '0.8B UD',
+    quant: 'Q4_K_XL',
+    size: '0.6 GB',
+    sizeGB: 0.6,
+    diskGB: 0.8,
+    context: '32k',
+    ctxMaxK: 32,
+    ctxPerGB: 0.004,
+    moe: false,
+    vision: false,
+    status: 'warm',
+    tags: ['Text'],
+    nodeCount: 1,
+    activitySummary: '0 requests seen · active 4 min ago'
+  },
+  {
+    name: 'Qwen3.5-2B',
+    fullId: 'Qwen3.5-2B-Q4_K_M',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 2,
+    paramsLabel: '2B',
+    quant: 'Q4_K_M',
+    size: '1.3 GB',
+    sizeGB: 1.3,
+    diskGB: 1.6,
+    context: '32k',
+    ctxMaxK: 32,
+    ctxPerGB: 0.008,
+    moe: false,
+    vision: false,
+    status: 'warm',
+    tags: ['Text'],
+    nodeCount: 1,
+    activitySummary: '0 requests seen · active 4 min ago'
+  },
+  {
+    name: 'Qwen3.5-4B-UD',
+    fullId: 'Qwen3.5-4B-UD-Q4_K_XL',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 4,
+    paramsLabel: '4B UD',
+    quant: 'Q4_K_XL',
+    size: '2.9 GB',
+    sizeGB: 2.9,
+    diskGB: 3.2,
+    context: '32k',
+    ctxMaxK: 32,
+    ctxPerGB: 0.012,
+    moe: false,
+    vision: false,
+    status: 'warm',
+    tags: ['Text'],
+    nodeCount: 1,
+    activitySummary: '0 requests seen · active 4 min ago'
+  },
+  {
+    name: 'Qwen3.6-27B-UD',
+    fullId: 'Qwen3.6-27B-UD-Q4_K_XL',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 27,
+    paramsLabel: '27B UD',
+    quant: 'Q4_K_XL',
+    size: '17.8 GB',
+    sizeGB: 17.8,
+    diskGB: 19,
+    context: '256k',
+    ctxMaxK: 256,
+    ctxPerGB: 0.025,
+    moe: false,
+    vision: false,
+    status: 'warm',
+    tags: ['Text'],
+    nodeCount: 1,
+    activitySummary: '0 requests seen · active 4 min ago'
+  },
+  {
+    name: 'Qwen3.6-35B-A3B-UD',
+    fullId: 'Qwen3.6-35B-A3B-UD-Q4_K_XL',
+    family: 'Qwen',
+    familyColor: 'family-2',
+    paramsB: 35,
+    paramsLabel: '35B A3B',
+    quant: 'Q4_K_XL',
+    size: '22.1 GB',
+    sizeGB: 22.1,
+    diskGB: 24,
+    context: '256k',
+    ctxMaxK: 256,
+    ctxPerGB: 0.026,
+    moe: true,
+    vision: false,
+    status: 'warm',
+    tags: ['Text'],
+    nodeCount: 1,
+    activitySummary: '0 requests seen · active 4 min ago'
+  }
+]
+export const MESH_NODES: MeshNode[] = [
+  {
+    id: 'self',
+    peerId: 'p1',
+    label: 'CARRACK',
+    subLabel: 'SERVING · YOU',
+    x: 58,
+    y: 52,
+    status: 'online',
+    role: 'self'
+  },
+  { id: 'lemony', peerId: 'p2', label: 'LEMONY-28', subLabel: 'SERVING', x: 36, y: 26, status: 'online' },
+  { id: 'lemony-29', peerId: 'p3', label: 'LEMONY-29', subLabel: 'SERVING', x: 28, y: 76, status: 'online' }
+]
